// src/app/modules/business-common/src/lib/components/passenger-types/enums/translation-keys.enum.ts
export enum TranslationKeys {
  PassengerTypes_KeyNode = 'PassengerTypes.',
  PassengerTypes_ADT = 'PassengerTypes.ADT',
  PassengerTypes_CHD = 'PassengerTypes.CHD',
  PassengerTypes_EXST = 'PassengerTypes.EXST',
  PassengerTypes_INF = 'PassengerTypes.INF',
  PassengerTypes_INS = 'PassengerTypes.INS',
  PassengerTypes_Plural_KeyNode = 'PassengerTypes.Plural_',
  PassengerTypes_Plural_ADT = 'PassengerTypes.Plural_ADT',
  PassengerTypes_Plural_CHD = 'PassengerTypes.Plural_CHD',
  PassengerTypes_Plural_INF = 'PassengerTypes.Plural_INF',
  PassengerTypes_Plural_TNG = 'PassengerTypes.Plural_TNG',
  PassengerTypes_TNG = 'PassengerTypes.TNG',
}